Improve preloading (Flickering WIP)
